refactor: 경로 수정 및 불필요한 코드 제거
